float

+++ Benchmark/spectralnorm/advanced/main.py
@@ -14,7 +14,7 @@ note, I don't see any concurrency in this code
 """
 
 import __static__
-from __static__ import int64, float64, CheckedList
+from __static__ import int64, CheckedList
 from typing import List, Tuple
 from typing import Callable as Function
 
@@ -23,15 +23,15 @@ def eval_A(i: float, j: float) -> float:
     return 1.0 / ((i + j) * (i + j + 1) // 2 + i + 1)
 
 
-def eval_times_u(func: Function[[Tuple[float64, List[float64]]], float64], u: List[float64]) -> List[float64]:
+def eval_times_u(func: Function[[Tuple[float, List[float]]], float], u: List[float]) -> List[float]:
     return [func((i, u)) for i in range(len(CheckedList[u]))]
 
 
-def eval_AtA_times_u(u: List[float64]) -> List[float64]:
+def eval_AtA_times_u(u: List[float]) -> List[float]:
     return eval_times_u(part_At_times_u, eval_times_u(part_A_times_u, u))
 
 
-def part_A_times_u(i_u: (float64, List[float64])) -> float64:
+def part_A_times_u(i_u: (float, List[float])) -> float:
     i, u = i_u
     partial_sum = 0
     for j, u_j in enumerate(u):
@@ -39,7 +39,7 @@ def part_A_times_u(i_u: (float64, List[float64])) -> float64:
     return partial_sum
 
 
-def part_At_times_u(i_u: (float64, List[float64])) -> float64:
+def part_At_times_u(i_u: (float, List[float])) -> float:
     i, u = i_u
     partial_sum = 0
     for j, u_j in enumerate(u):
